Implement GICv3 and bug fixes in D-cached memory area
Implemented GICv3 support and lots of bug fixed in Virtual Memory mappings
interms of Cache flags, previously it was using device memory mair attribute
for normal memory maps accidentally, and now it is fixed.

Manas Kamal Choudhury

+++ BaseHdr/Mm/mmfile.h
@@ -35,10 +35,11 @@
 #include <stdint.h>
 #include <Fs/vfs.h>
 
+
 typedef struct __page_cache__ {
 	uint64_t physicalPage;
 	uint64_t diskBlock;
-	uint32_t pageIndex;
+	uint64_t pageIndex;
 	struct __page_cache__* next;
 	struct __page_cache__* prev;
 }AuMMPageCache;
